move run event mirror into daemon module
